requirements.txt
appdirs==1.4.4
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.9.3
black==23.7.0
blinker==1.6.2
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.6
colorama==0.4.4
constantly==15.1.0
cryptography==40.0.2
cssselect==1.2.0
EasyProcess==0.3
exceptiongroup==1.1.2
filelock==3.12.0
flake8==3.8.4
flask==2.3.3
fuzzywuzzy==0.18.0
greenlet==2.0.2
h11==0.14.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
incremental==22.10.0
isort==5.12.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
loguru==0.5.3
lxml==4.9.2
MarkupSafe==2.1.3
mccabe==0.6.1
mypy-extensions==1.0.0
mysqlclient==2.2.0
numpy==1.24.4
outcome==1.2.0
packaging==23.1
pandas==2.0.3
parsel==1.8.1
pathspec==0.11.1
platformdirs==3.9.1
playwright==1.37.0
Protego==0.2.1
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.6.0
pycparser==2.21
PyDispatcher==2.0.7
pyee==9.0.4
pyflakes==2.2.0
pyOpenSSL==23.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.15.0
pytz==2023.3.post1
PyVirtualDisplay==1.3.2
PyYAML==6.0.1
queuelib==1.6.2
regex==2020.11.13
requests==2.29.0
requests-file==1.5.1
Scrapy==2.8.0
scrapy-playwright==0.0.32
selenium==4.9.1
service-identity==21.1.0
ShopifyScraper==0.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.1
tldextract==3.4.1
toml==0.10.2
tomli==2.0.1
trio==0.22.2
trio-websocket==0.10.3
Twisted==22.10.0
typed-ast==1.5.5
typing-extensions==4.5.0
tzdata==2023.3
undetected-chromedriver==3.5.0
urllib3==1.26.15
w3lib==2.1.1
websockets==11.0.3
werkzeug==2.3.7
win32-setctime==1.0.3
wsproto==1.2.0
zipp==3.17.0
zope.interface==6.0
